requirements.txt
streamlit
pandas
matplotlib
openpyxl
python-docx==0.8.11
pdfplumber
easyocr
Pillow
together
replicate
PyPDF2==3.0.1